+++ ui/src/pages/MainPage.vue
@@ -139,18 +139,24 @@ watch(() => [app.model.args.numerators, app.model.args.denominator], (_) => {
             </template>
           </PlNumberField>
           <PlNumberField
-            v-model="app.model.args.pAdjFCThreshold"
+            v-model="app.model.args.pAdjThreshold"
             label="Adjusted p-value" :minValue="0" :maxValue="1" :step="0.01"
           />
         </PlRow>
         <!-- Add warnings if selected threshold are out of most commonly used bounds -->
-        <PlAlert v-if="app.model.args.pAdjFCThreshold > 0.05" type="warn">
+        <PlAlert v-if="app.model.args.pAdjThreshold > 0.05" type="warn">
           {{ "Warning: The selected adjusted p-value threshold is higher than the most commonly used 0.05" }}
         </PlAlert>
         <PlAlert v-if="app.model.args.log2FCThreshold < 0.6" type="warn">
           {{ "Warning: The selected Log2(FC) threshold may be too low for most use cases" }}
         </PlAlert>
       </PlAccordionSection>
+      <PlAlert v-if="app.model.outputs.emptyCheck == 'notFullRank'" type="error">
+        {{ "Error: The model matrix is not full rank, so the model cannot be fit as specified. \
+              One or more variables or interaction terms in the design formula are linear\
+              combinations of the others and must be removed. Please, check the metadata \
+              columns included in the Design section." }}
+      </PlAlert>
     </PlSlideModal>
   </PlBlockPage>
 </template>