+++ software/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @platforma-open/milaboratories.run-deseq2-r.software
 
+## 1.7.0
+
+### Minor Changes
+
+- db81af6: Added conditional execution step if model matrix not full rank
+
 ## 1.6.0
 
 ### Minor Changes